+++ .eslintrc.js
@@ -3,11 +3,11 @@ module.exports = {
     browser: true,
     es2021: true
   },
-  extends: [
-    'eslint:recommended',
-    'plugin:node/recommended',
-    'plugin:@typescript-eslint/recommended'
-  ],
+  // extends: [
+  //   'eslint:recommended',
+  //   'plugin:node/recommended',
+  //   'plugin:@typescript-eslint/recommended'
+  // ],
   parser: '@typescript-eslint/parser',
   parserOptions: {
     ecmaVersion: 'latest',