+++ scripts/verifyCommits.ts
@@ -1,13 +1,9 @@
-// const chalk = require('chalk')
 // use 'picocolors' to replace 'chalk'
-const colors = require('picocolors')
+import colors from 'picocolors'
+import { readFileSync } from 'fs'
 
-// const msgPath = process.env.GIT_PARAMS
 const msgPath = process.argv[2]
-const msg = require('fs').readFileSync(msgPath, 'utf-8').trim()
-
-console.log()
-console.log(colors.bgBlue('[sure-lint] Commit Msg: '), msg)
+const msg = readFileSync(msgPath, 'utf-8').trim()
 
 const commitRE =
   /^(revert: )?(feat|fix|docs|dx|style|refactor|perf|test|workflow|build|ci|chore|types|wip|release)(\(.+\))?: .{1,50}/
@@ -21,10 +17,8 @@ if (!commitRE.test(msg)) {
       colors.red(
         `  Proper commit message format is required for automated changelog generation. Examples:\n\n`
       ) +
-      `    ${colors.green(`feat(compiler): add 'comments' option`)}\n` +
-      `    ${colors.green(
-        `fix(input): handle events on blur (close #28)`
-      )}\n\n` +
+      `    ${colors.green(`feat: add 'comments' option`)}\n` +
+      `    ${colors.green(`fix: handle events on blur (close #25)`)}\n\n` +
       colors.red(`  See .github/commit-convention.md for more details.\n`)
   )
   process.exit(1)