+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+# 1.0.0 (2022-06-24)
+
+
+### Bug Fixes
+
+* replace 'picocolors' with 'chalk' ([6f4ecb3](https://github.com/suressk/sure-lint/commit/6f4ecb3857671cc71adaaf327c7496f5e6c31d83))
+* **verifyCommits.js:** add a blank line after the tip of commit-msg ([a4b64d2](https://github.com/suressk/sure-lint/commit/a4b64d256a40f4685fbaa79168b264b97c4740eb))
+
+
+### Features
+
+* add .prettierignore ([49218a0](https://github.com/suressk/sure-lint/commit/49218a0db247544dfdf3558d35c3257ced4b3da9))
+* add eslint-plugin-node ([e025e3c](https://github.com/suressk/sure-lint/commit/e025e3cfb78b44712b7e713296fdb7487ed8dc18))
+* add readme and license ([7ac113c](https://github.com/suressk/sure-lint/commit/7ac113ca53259f5d320d35f9bcd78a669d51126b))
+* **changelog:** add changelog ([812e912](https://github.com/suressk/sure-lint/commit/812e91296e34b694dc43afeb94d23a08eff6d3e8))
+
+
+
 # 1.0.0 (2022-06-21)